Add Jest and reorganize helpers within test folder
