+++ tests/helpers/get-prop.js
