+++ tests/helpers/get-rendered-prop.js
