+++ tests/integration/posts.js
@@ -8,15 +8,15 @@ const SUCCESS = 'success';
 // actually run.
 chai.use( require( 'chai-as-promised' ) );
 const expect = chai.expect;
-const httpTestUtils = require( './helpers/http-test-utils' );
+const httpTestUtils = require( '../helpers/http-test-utils' );
 
 const WPAPI = require( '../../' );
 const WPRequest = require( '../../lib/constructors/wp-request.js' );
 
 // Inspecting the titles of the returned posts arrays is an easy way to
 // validate that the right page of results was returned
-const getTitles = require( './helpers/get-rendered-prop' ).bind( null, 'title' );
-const credentials = require( './helpers/constants' ).credentials;
+const getTitles = require( '../helpers/get-rendered-prop' ).bind( null, 'title' );
+const credentials = require( '../helpers/constants' ).credentials;
 
 // Define some arrays to use ensuring the returned data is what we expect
 // it to be (e.g. an array of the titles from posts on the first page)
